Delete json parser, delete userController
