+++ DevelopGitHub/GitHubExtension/GitHubExtension.Security.WebApi.Library/Controllers/RepositoryController.cs
@@ -8,6 +8,8 @@ using GitHubExtension.Security.DAL.Interfaces;
 using GitHubExtension.Security.WebApi.Library.Mappers;
 using GitHubExtension.Security.WebApi.Library.Services;
 using Microsoft.AspNet.Identity;
+using GitHubExtension.Security.StorageModels.Identity;
+using GitHubExtension.Security.DAL.Infrastructure;
 
 namespace GitHubExtension.Security.WebApi.Library.Controllers
 {
@@ -15,11 +17,14 @@ namespace GitHubExtension.Security.WebApi.Library.Controllers
     {
         private IGithubService _guGithubService;
         private readonly ISecurityContext _securityContext;
+        private readonly ApplicationUserManager _userManager;
 
-        public RepositoryController(IGithubService guGithubService, ISecurityContext securityContext)
+        public RepositoryController(IGithubService guGithubService, ISecurityContext securityContext,
+            ApplicationUserManager userManager)
         {
             _guGithubService = guGithubService;
             _securityContext = securityContext;
+            _userManager = userManager;
         }
 
         // GET: api/repos/:id
@@ -60,6 +65,39 @@ namespace GitHubExtension.Security.WebApi.Library.Controllers
             var gitHubCollaborators = await _guGithubService.GetCollaboratorsForRepo(userName, repoName, token);
 
             return Ok(gitHubCollaborators);
+        }
+
+        [Authorize]
+        [Route("api/repos/current")]
+        [HttpPatch]
+        public async Task<IHttpActionResult> UpdateProject(Repository repo)
+        {
+            string currentUserId = User.Identity.GetUserId();
+            User user = await _userManager.Users.FirstOrDefaultAsync(x => x.Id == currentUserId);
+            if (user == null)
+                return NotFound();
+
+            var repository = _securityContext.UserRepository.FirstOrDefault(r => r.RepositoryId == repo.Id && r.UserId == user.Id);
+            if (repository == null)
+            {
+                ModelState.AddModelError("repo", string.Format("user with id '{0}' does not have a repository with id '{1}'", user.Id, repo.Id));
+                return BadRequest(ModelState);
+            }
+
+            var claimsIdentity = await user.GenerateUserIdentityAsync(_userManager, DefaultAuthenticationTypes.ApplicationCookie);
+            var existingClaim = claimsIdentity.Claims.FirstOrDefault(c => c.Type == "CurrentProjectId");
+            if (existingClaim != null)
+                _userManager.RemoveClaim(user.Id, existingClaim);
+
+            var addClaimResult = await _userManager.AddClaimAsync(user.Id, new Claim("CurrentProjectId", repo.Id.ToString()));
+
+            if (!addClaimResult.Succeeded)
+            {
+                ModelState.AddModelError("CurrentpProject", "Failed to update current project");
+                return BadRequest(ModelState);
+            }
+
+            return Ok();
         }
     }
 }
